+++ app/src/main/java/com/muhammetkdr/weatherapp/model/forecast/Clouds.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 
 data class Clouds(
     @SerializedName("all")
-    val all: Int?
+    val all: Double?
 )
