+++ app/src/main/java/com/muhammetkdr/weatherapp/model/forecast/Coord.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Coord(
     @SerializedName("lat")
-    val lat: Int?,
+    val lat: Double?,
     @SerializedName("lon")
-    val lon: Int?
+    val lon: Double?
 )
