+++ app/src/main/java/com/muhammetkdr/weatherapp/model/forecast/Wind.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Wind(
     @SerializedName("deg")
-    val deg: Int?,
+    val deg: Double?,
     @SerializedName("gust")
     val gust: Double?,
     @SerializedName("speed")
